move ingredient icons to components and callbacks
